refactor inputarea own component
