+++ client/src/components/organisms/RightSection/RightSection.js
@@ -12,7 +12,7 @@ import {
   setActivePanelChat,
   setShouldFocusChatInput,
 } from '../../../redux/userSlice';
-import { ArrowForwardIcon, ChevronRightIcon } from '@chakra-ui/icons';
+import { ChevronRightIcon } from '@chakra-ui/icons';
 import {
   createChatWithSelectedDropdownCourse,
   fetchChat,
@@ -24,56 +24,11 @@ import {
 } from '../../../redux/uiSlice';
 import SearchBarInput from '../../molecules/SearchBar/SearchBarInput';
 import { HotKeys } from 'react-hotkeys';
+import InputArea from '../../atoms/InputArea/InputArea';
 
 const keyMap = {
   SHOW_SEARCH: 'command+f',
 };
-
-const Message = ({ value }) => (
-  <div className={styles.message}>
-    <p>{value}</p>
-  </div>
-);
-
-const InputArea = ({
-  currentUserInput,
-  setInputText,
-  onInputSubmit,
-  inputRef,
-  disableInput,
-}) => (
-  <div className={styles.inputSection}>
-    <div className={styles.inputArea}>
-      <input
-        className={styles.input}
-        placeholder="Enter a prompt..."
-        value={currentUserInput || ''}
-        onChange={e => setInputText(e.target.value)}
-        onKeyDown={onInputSubmit}
-        ref={inputRef}
-      />
-      <button
-        className={styles.sendBtn}
-        onClick={e => {
-          if (currentUserInput.trim()) onInputSubmit(e);
-        }}
-        style={
-          currentUserInput.trim() && !disableInput
-            ? {}
-            : { cursor: 'not-allowed', opacity: 0.5 }
-        }
-        disabled={disableInput}
-      >
-        <ArrowForwardIcon />
-      </button>
-    </div>
-    <Message
-      value={
-        'CourseGPT may produce inaccurate information about instructors or course content. [CourseGPT 2023 Version]'
-      }
-    />
-  </div>
-);
 
 const RightSection = () => {
   const dispatch = useDispatch();
