+++ client/src/components/organisms/RightSection/RightSection.module.css
@@ -16,60 +16,6 @@
   }
 }
 
-.inputSection {
-  position: absolute;
-  bottom: 0;
-  width: 100%;
-  height: 12vh;
-  background: linear-gradient(
-    0deg,
-    rgba(53, 55, 68, 1) 23%,
-    rgba(53, 55, 68, 0.8239889705882353) 45%,
-    rgba(53, 55, 68, 0.6951374299719888) 59%,
-    rgba(53, 55, 68, 0) 100%
-  );
-}
-
-.inputArea {
-  width: 60%;
-  text-align: center;
-  margin: auto;
-  position: relative;
-}
-
-.input {
-  padding: 12px;
-  box-shadow: rgba(48, 45, 45, 0.2) 0px 7px 29px 0px;
-  border-radius: 4px;
-  font-size: medium;
-  border: none;
-  width: 100%;
-  outline: none;
-  background-color: var(--gray2);
-  color: white;
-}
-
-.input::placeholder {
-  color: var(--placeholder);
-}
-
-.message {
-  font-size: 12px;
-  margin-top: 10px;
-  text-align: center;
-  color: rgb(197, 197, 210);
-}
-
-.sendBtn {
-  background: transparent;
-  position: absolute;
-  right: 0;
-  outline: none;
-  border: none;
-  height: 48px;
-  width: 66px;
-}
-
 .toggleSidepanelBtn {
   position: absolute;
   top: 10px;
